Remove underscore from widget compoenent Calendar for consistent naming
